% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{Dist}
\alias{Dist}
\title{NGBoost distributions}
\usage{
Dist(
  dist = c("Bernoulli", "k_categorical", "StudentT", "Poisson", "Laplace", "Cauchy",
    "Exponential", "LogNormal", "MultivariateNormal", "Normal"),
  K
)
}
\arguments{
\item{dist}{NGBoost distributions
\itemize{
\item Bernoulli
\item k_categorical
\item StudentT
\item Poisson
\item Laplace
\item Cauchy
\item Exponential
\item LogNormal
\item MultivariateNormal
\item Normal
}}

\item{K}{Used only with k_categorical and MultivariateNormal}
}
\description{
NGBoost distributions
}
