% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngb_dist_class.R
\name{NGBDistClass}
\alias{NGBDistClass}
\title{NGBoost classifier distribution object}
\value{
A NGBDist Class
}
\description{
NGBoost classifier distribution object

NGBoost classifier distribution object
}
\author{
Resul Akay
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NGBDistClass-new}{\code{NGBDistClass$new()}}
\item \href{#method-NGBDistClass-class_probs}{\code{NGBDistClass$class_probs()}}
\item \href{#method-NGBDistClass-clone}{\code{NGBDistClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistClass-new}{}}}
\subsection{Method \code{new()}}{
Initialize the NGBoost distribution object.
It is initialize internally.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistClass$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A python ngboost.distns.distn.Distn.uncensor object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistClass-class_probs"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistClass-class_probs}{}}}
\subsection{Method \code{class_probs()}}{
Class Probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistClass$class_probs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
