% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngb_dist_reg.R
\name{NGBDistReg}
\alias{NGBDistReg}
\title{NGBoost regression distribution object}
\value{
A NGBDist Class
}
\description{
NGBoost regression distribution object

NGBoost regression distribution object
}
\author{
Resul Akay
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NGBDistReg-new}{\code{NGBDistReg$new()}}
\item \href{#method-NGBDistReg-rvs}{\code{NGBDistReg$rvs()}}
\item \href{#method-NGBDistReg-pdf}{\code{NGBDistReg$pdf()}}
\item \href{#method-NGBDistReg-logpdf}{\code{NGBDistReg$logpdf()}}
\item \href{#method-NGBDistReg-cdf}{\code{NGBDistReg$cdf()}}
\item \href{#method-NGBDistReg-logcdf}{\code{NGBDistReg$logcdf()}}
\item \href{#method-NGBDistReg-sf}{\code{NGBDistReg$sf()}}
\item \href{#method-NGBDistReg-logsf}{\code{NGBDistReg$logsf()}}
\item \href{#method-NGBDistReg-ppf}{\code{NGBDistReg$ppf()}}
\item \href{#method-NGBDistReg-isf}{\code{NGBDistReg$isf()}}
\item \href{#method-NGBDistReg-entropy}{\code{NGBDistReg$entropy()}}
\item \href{#method-NGBDistReg-interval}{\code{NGBDistReg$interval()}}
\item \href{#method-NGBDistReg-mean}{\code{NGBDistReg$mean()}}
\item \href{#method-NGBDistReg-median}{\code{NGBDistReg$median()}}
\item \href{#method-NGBDistReg-std}{\code{NGBDistReg$std()}}
\item \href{#method-NGBDistReg-var}{\code{NGBDistReg$var()}}
\item \href{#method-NGBDistReg-clone}{\code{NGBDistReg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-new"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-new}{}}}
\subsection{Method \code{new()}}{
Initialize the NGBoost distribution object.
It is initialize internally.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A python ngboost.distns.distn.Distn.uncensor object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-rvs"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-rvs}{}}}
\subsection{Method \code{rvs()}}{
Random variates of given type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$rvs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Random variates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-pdf}{}}}
\subsection{Method \code{pdf()}}{
Probability density function at x of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$pdf(quantiles, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Probability density function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-logpdf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-logpdf}{}}}
\subsection{Method \code{logpdf()}}{
Log of the probability density function at x of the given
RV.
This uses a more numerically accurate calculation if available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$logpdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log of the probability density function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-cdf}{}}}
\subsection{Method \code{cdf()}}{
Cumulative distribution function of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$cdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cumulative distribution function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-logcdf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-logcdf}{}}}
\subsection{Method \code{logcdf()}}{
Log of the cumulative distribution function at x of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$logcdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log of the cumulative distribution function evaluated at
quantiles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-sf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-sf}{}}}
\subsection{Method \code{sf()}}{
Survival function (1 - 'cdf') at x of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$sf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Survival function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-logsf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-logsf}{}}}
\subsection{Method \code{logsf()}}{
Log of the survival function of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$logsf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}

\item{\code{return}}{Log of the survival function evaluated at quantiles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-ppf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-ppf}{}}}
\subsection{Method \code{ppf()}}{
Percent point function (inverse of 'cdf') at q of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$ppf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{array, Lower tail probability.}

\item{\code{...}}{Other arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
quantile corresponding to the lower tail probability q
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-isf"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-isf}{}}}
\subsection{Method \code{isf()}}{
Inverse survival function (inverse of sf) at q of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$isf(q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{array, Lower tail probability.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Quantile corresponding to the upper tail probability, q.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-entropy"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-entropy}{}}}
\subsection{Method \code{entropy()}}{
Differential entropy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$entropy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Differential entropy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-interval"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-interval}{}}}
\subsection{Method \code{interval()}}{
Confidence interval with equal areas around the median.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$interval(confidence)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{confidence}}{Probability that an rv will be drawn from the returned
range.
Each value should be in the range \code{0>= confidence <=1}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nd-points of range that contain \code{100 * confidence} of
the rv's possible values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-mean"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-mean}{}}}
\subsection{Method \code{mean()}}{
Mean of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mean of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-median"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-median}{}}}
\subsection{Method \code{median()}}{
Median of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$median()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The median of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-std"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-std}{}}}
\subsection{Method \code{std()}}{
Standard deviation of the distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$std()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The standard deviation of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-var"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-var}{}}}
\subsection{Method \code{var()}}{
Variance of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$var()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The variance of the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NGBDistReg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NGBDistReg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NGBDistReg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
