% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearners.R
\name{Ridge}
\alias{Ridge}
\title{Linear least squares with l2 regularization.}
\usage{
Ridge(alpha = 0, random_state = NULL)
}
\arguments{
\item{alpha}{Float, ndarray of shape (n_targets,).
Regularization strength; must be a positive float. Regularization improves
the conditioning of the problem and reduces the variance of the estimates.
Larger values specify stronger regularization.}

\item{random_state}{sklearn.linear_model._ridge.Ridge}
}
\value{
sklearn.linear_model._ridge.Ridge object
}
\description{
Linear least squares with l2 regularization.
}
\author{
Resul Akay
}
