% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngb_dist.R
\name{NgbDist}
\alias{NgbDist}
\title{NGBoost distribution object}
\value{
A NgbDist Class
}
\description{
NGBoost distribution object

NGBoost distribution object
}
\author{
Resul Akay
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NgbDist$new()}}
\item \href{#method-rvs}{\code{NgbDist$rvs()}}
\item \href{#method-pdf}{\code{NgbDist$pdf()}}
\item \href{#method-logpdf}{\code{NgbDist$logpdf()}}
\item \href{#method-cdf}{\code{NgbDist$cdf()}}
\item \href{#method-logcdf}{\code{NgbDist$logcdf()}}
\item \href{#method-sf}{\code{NgbDist$sf()}}
\item \href{#method-logsf}{\code{NgbDist$logsf()}}
\item \href{#method-ppf}{\code{NgbDist$ppf()}}
\item \href{#method-isf}{\code{NgbDist$isf()}}
\item \href{#method-entropy}{\code{NgbDist$entropy()}}
\item \href{#method-interval}{\code{NgbDist$interval()}}
\item \href{#method-mean}{\code{NgbDist$mean()}}
\item \href{#method-median}{\code{NgbDist$median()}}
\item \href{#method-std}{\code{NgbDist$std()}}
\item \href{#method-var}{\code{NgbDist$var()}}
\item \href{#method-clone}{\code{NgbDist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the NGBoost distribution object.
It is initialize internally.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A python ngboost.distns.distn.Distn.uncensor object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rvs"></a>}}
\if{latex}{\out{\hypertarget{method-rvs}{}}}
\subsection{Method \code{rvs()}}{
Random variates of given type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$rvs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Random variates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-pdf}{}}}
\subsection{Method \code{pdf()}}{
Probability density function at x of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$pdf(quantiles, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Probability density function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logpdf"></a>}}
\if{latex}{\out{\hypertarget{method-logpdf}{}}}
\subsection{Method \code{logpdf()}}{
Log of the probability density function at x of the given RV.
This uses a more numerically accurate calculation if available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$logpdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log of the probability density function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-cdf}{}}}
\subsection{Method \code{cdf()}}{
Cumulative distribution function of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$cdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cumulative distribution function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logcdf"></a>}}
\if{latex}{\out{\hypertarget{method-logcdf}{}}}
\subsection{Method \code{logcdf()}}{
Log of the cumulative distribution function at x of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$logcdf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log of the cumulative distribution function evaluated at
quantiles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sf"></a>}}
\if{latex}{\out{\hypertarget{method-sf}{}}}
\subsection{Method \code{sf()}}{
Survival function (1 - 'cdf') at x of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$sf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Survival function evaluated at quantiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logsf"></a>}}
\if{latex}{\out{\hypertarget{method-logsf}{}}}
\subsection{Method \code{logsf()}}{
Log of the survival function of the given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$logsf(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{The quantiles}

\item{\code{return}}{Log of the survival function evaluated at quantiles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ppf"></a>}}
\if{latex}{\out{\hypertarget{method-ppf}{}}}
\subsection{Method \code{ppf()}}{
Percent point function (inverse of 'cdf') at q of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$ppf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{array, Lower tail probability.}

\item{\code{...}}{Other arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
quantile corresponding to the lower tail probability q
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isf"></a>}}
\if{latex}{\out{\hypertarget{method-isf}{}}}
\subsection{Method \code{isf()}}{
Inverse survival function (inverse of sf) at q of the
given RV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$isf(q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{array, Lower tail probability.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Quantile corresponding to the upper tail probability, q.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-entropy"></a>}}
\if{latex}{\out{\hypertarget{method-entropy}{}}}
\subsection{Method \code{entropy()}}{
Differential entropy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$entropy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Differential entropy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interval"></a>}}
\if{latex}{\out{\hypertarget{method-interval}{}}}
\subsection{Method \code{interval()}}{
Confidence interval with equal areas around the median.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$interval(confidence)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{confidence}}{Probability that an rv will be drawn from the returned range.
Each value should be in the range \code{0>= confidence <=1}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nd-points of range that contain \code{100 * confidence} of
the rv's possible values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean"></a>}}
\if{latex}{\out{\hypertarget{method-mean}{}}}
\subsection{Method \code{mean()}}{
Mean of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mean of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-median"></a>}}
\if{latex}{\out{\hypertarget{method-median}{}}}
\subsection{Method \code{median()}}{
Median of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$median()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The median of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-std"></a>}}
\if{latex}{\out{\hypertarget{method-std}{}}}
\subsection{Method \code{std()}}{
Standard deviation of the distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$std()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The standard deviation of the distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-var"></a>}}
\if{latex}{\out{\hypertarget{method-var}{}}}
\subsection{Method \code{var()}}{
Variance of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$var()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The variance of the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgbDist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
