% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ngboost}
\alias{ngboost}
\title{R interface for NGBoost package}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ngboost
}
\description{
The goal of ngboost is to provide an R interface for the Python package \href{https://stanfordmlgroup.github.io}{NGBoost}
}
\keyword{datasets}
